fix(utils): invert MSDB indexing
Instead of counting from left (MSB) to right (LSB) do it from right to left so
the results will be independent of the size of the integers that code the node
indexes. For example,

* 63 -> 0
* 62 -> 1
* 61 -> 2

This change in the indexes caused some tests to fail (besides the MSDB one which
was expected), mainly because we were representing sets as vectors (so the same
results reordered failed). Change `Vec` to `HashSet` for partitions used in
Valiant.

NOTE: There is a particular test, `test_valiant_reduce`, that as a result of
this reordering returned a smaller `S` which sounds odd since `valiant_basic`
already orders the partitions to use the smallest ones first.